Refactor Instruments component file structure.
Moved Instruments component and updated imports to align with the new directory structure. This enhances code organization and simplifies the pathing for related files.
