+++ cypress/component/Instruments.cy.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Instruments from '../../src/Instruments';
+import Instruments from '../../src/pages/Instruments';
 import { mount } from '@cypress/react';
 
 describe('<Instruments />', () => {
